refactor: make a ui package
The ui object will get huge since I'm about to introduce a couple new
screens. This will just allow for me to dump them all in one package.
